+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+
+- @FXML delegate (https://github.com/edvin/tornadofx/issues/34)
+- i18n Support (https://github.com/edvin/tornadofx/issues/29)
+- `TableView.column()` support for `ObservableValue` member fields
+- `FX.runAndWait` sync execution helper
+- `TableColumn.makeEditable` extension
+- `JsonModelAuto` automatically converts from JavaFX Properties to JSON
+
+### Changed
+
+- `Fragment` should not be injectable (https://github.com/edvin/tornadofx/issues/31)
+- `FXTableView` removed, `column` functions are now extensions on `TableView`
+
+## [1.3.1] - 2016-02-28
+
+### Added
+- Arbitrary properties per `Component` (https://github.com/edvin/tornadofx/issues/23)
 - `singleAssign()` property delegates (https://github.com/edvin/tornadofx/issues/17)
 - Tests for `singleAssign()`
 - BorderPane builder (https://github.com/edvin/tornadofx/pull/16)
@@ -18,6 +35,7 @@ All notable changes to this project will be documented in this file.
 - Builder for Tab now require the content pane as input parameter to avoid confusion (https://github.com/edvin/tornadofx/issues/8)
 - UIComponent#openModal() no longer requests focus for root component - caller can decide instead
 - Property delegate now produces non-nullable property types
+- `GridPane.row()` no long hogs the GridPane `userData` object to track rowId
 
 ## [1.3.0] - 2016-02-19